Refactor dangerzone to move GUI code into its own module
