+++ dangerzone/gui/doc_selection_widget.py
