+++ dev_scripts/dangerzone-cli
@@ -1,0 +1,1 @@
+dangerzone
